add new class to put apart all the Fi-ware specific code
