+++ printimages.py
@@ -25,10 +25,10 @@
 author = 'jmpr22'
 import sys
 import os
-import glancesync
+import glancesync_fi
 
 if __name__ == '__main__':
-    sync_obj = glancesync.GlanceSync()
+    sync_obj = glancesync_fi.GlanceSyncFi()
     if len(sys.argv) > 1:
         regions = sys.argv[1:]
     else:
