+++ sync.py
@@ -26,7 +26,7 @@ author = 'jmpr22'
 import sys
 import os
 import os.path
-from glancesync import GlanceSync
+from glancesync_fi import GlanceSyncFi as GlanceSync
 
 if __name__ == '__main__':
     glancesync = GlanceSync()
